requirements.txt
amqp==5.0.3
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
boto3==1.16.59
botocore==1.19.59
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-cors-headers==3.6.0
django-cuser==2017.3.16
django-filter==2.4.0
django-modeltranslation==0.16.1
django-otp==1.0.2
django-phonenumber-field==5.0.0
django-rest-swagger==2.2.0
django-silk==4.1.0
django-storages==1.11.1
django-tinymce==3.2.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gprof2dot==2019.11.30
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kavenegar==1.1.2
kombu==5.0.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
phonenumbers==8.12.15
Pillow==8.0.1
prompt-toolkit==3.0.14
psycopg2-binary==2.8.5
pycodestyle==2.6.0
Pygments==2.7.3
PyJWT==2.0.0
pylint==2.6.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2020.4
redis==3.5.3
requests==2.25.1
s3transfer==0.3.4
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.4.0
